CUDA bench results (1x GB200) + bf16 bench support + nickname-scope cleanup
docs/model_eval_throughput.md: the specced cluster experiment RAN
(2026-07-19, 1x NVIDIA GB200 as reported by torch, owner-extended to bf16
and a 4-GPU replica-sharding leg). New CUDA results section: full grids for
both checkpoints (emeta fp32 ~15k evals/s, fp16/bf16 ~60k; m50 fp32 ~3.7k,
fp16/bf16 ~26-27k — tensor cores compress the emeta:m50 ratio from 4.1x
fp32 to 2.2x bf16), partial-batch study (NO cliff on CUDA: cold==warm,
proportional to rows — pad-to-batch is an MPS-only workaround), 4-way
replica leg (99-100% scaling at batch 256), CPU crossover on cluster cores
(~0.4x the laptop CPU: GPU dominates 97-369x), decision synthesis and an
updated recommendation (emeta default survives in parity-clean fp32 at
4.1x; bf16 narrows to 2.2x pending any reduced-precision parity story;
fp8 noted as future probe, not built).

scripts/bench_model_eval.py: bf16 as a first-class dtype (real bfloat16
copy traced to TorchScript, CUDA-oriented; skipped on cpu like fp16).

src/pokezero/neural_policy.py: two dtype-following casts so real fp16/bf16
weight copies can run — _embed_expanded_inputs no longer hard-upcasts
numeric features to fp32, _masked_mean weights masks in the values dtype.
Bit-identical under fp32 (full neural/export suites pass; the two
test_neural_selfplay promote-gate failures pre-exist on main).

docs/crate_search_design.md, docs/leaf_observation_column_map.md: nickname
idents marked OUT OF SCOPE (owner decision 2026-07-19, randbats-only
project) instead of a deferred revisit.

+++ src/pokezero/neural_policy.py
@@ -1093,7 +1093,13 @@ if nn is not None:  # pragma: no cover - optional dependency path.
             token_embeddings = self.token_type_embedding(
                 token_type_ids.clamp(min=0, max=self.config.token_type_vocab_size - 1).long()
             )
-            numeric_embeddings = self.numeric_projection(numeric_features.float())
+            # Match the projection weight dtype rather than hard-upcasting to
+            # fp32: identical under fp32 weights, and keeps real fp16/bf16
+            # weight copies traceable (fp32 activations against half weights
+            # hard-error in F.linear).
+            numeric_embeddings = self.numeric_projection(
+                numeric_features.to(self.numeric_projection.weight.dtype)
+            )
             return category_embeddings + token_embeddings + numeric_embeddings
 
         def _pool_encoded_history(self, *, encoded: Any, attention_mask: Any, history_mask: Any) -> Any:
@@ -3058,7 +3064,10 @@ def _validate_row_indexed_tensor_shapes(
 
 def _masked_mean(values: Any, mask: Any) -> Any:
     torch_module = require_torch()
-    weights = mask.float().unsqueeze(-1)
+    # Weight in the values dtype: identical under fp32, and avoids promoting
+    # reduced-precision (fp16/bf16) activations to fp32 mid-forward, which
+    # would dtype-mismatch the downstream heads on a real half/bf16 copy.
+    weights = mask.to(values.dtype).unsqueeze(-1)
     denominator = weights.sum(dim=1).clamp(min=1.0)
     return (values * weights).sum(dim=1) / denominator
